pyproject.toml
[tool.poetry]
name = "auth-class"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
